lettuce: QCamera2: Change maxFps type as float for display metadata
Issue:
Display is getting the refresh rate as 0 fps for camera due to
mismatch in type of maxFps value. Display is expecting refresh rate
metadata as float type but camera is setting maxFps as integer type
which is causing the issue at display side.

Fix:
Update metdata with maxFps type as float from camera

Change-Id: I2682d9c9b539cbc1bf002575a68fe47969ec0fab
Signed-off-by: ghostrider-reborn <adithya.r02@outlook.com>
Signed-off-by: Ad!thya R <gh0strider.2k18.reborn@gmail.com>